Remove boilerplate for implementing `EventRetrieving` for contracts (gnosis/gp-v2-services#619)
This PR was re-created from #614 because there were issues on xDAI. The issues were unrelated to the `EventRetrieving` trait and instead caused by issues in the `contracts/build.rs` refactor. That refactor has been reverted in this PR and will be proposed in a follow up change.

This PR addresses comments from the previous PR, specifically:
- Implements a declarative macro to implement `EventRetrieving` for contracts bindings generated by `ethcontract`.
- Removes the `web3()` function from the `EventRetrieving` trait, requiring a `BlockRetrieving` to be passed when constructing event handler

### Test Plan

No logic changes, just refactoring (and moveing small amounts of code around). CI should continue to pass.