Single Order Solver abstraction (gnosis/gp-v2-services#697)
Preparing for the Paraswap solver, I noticed that a common abstraction will likely be "sequential" solvers that solve the entire set of liquidity individually (e.g Matcha, Balancer SOR, Paraswap, 1Inch).

This PR therefore extracts the split logic into a separate component. In the future this component could also do logic like time management (extrapolating how many more orders can be attempted).

We might also want to pass the remaining liquidity to the SingleOrderSolving trait, however since it's not yet needed we can cross that bridge later.

Also note, that we are now fetching available protocol on every fetch. If this is an issue, we could either start caching the protocols, or extend the trait with some `prepare` method that can return an associated `Context` type which we would pass in into each individual run of the SingleOrderSolving.

### Test Plan
CI (right now this component is too dumb to justify unit tests, as the only logic - extract limit orders from liquidity - is enforced by the compiler.