Support unknown tokens in Paraswap Solver (gnosis/gp-v2-services#743)
For "unknown" tokens to work with paraswap we need to add `fromDecimals` and `toDecimals` to the query (cf. https://developers.paraswap.network/api/get-rate-for-a-token-pair). I had already added the fields to the struct but forgot to populate them on the price query.

This lead to some errors executing the price query in staging (and thus paraswap not being chosen).

### Test Plan
Adjusted the e2e test to trade OWL (which is unknown to Paraswap)