Add allowance manager component (gnosis/gp-v2-services#786)
This PR adds an allowance manager component that can be used to generate approval interactions based on allowances. The allowance manager acts similarly to how allowances are handled in the Uniswap liquidity/settlement handing, in that allowances are pre-fetched and then later used to generate the required interactions.

Additionally, a method for fetching a single approval was also added that can be used by things like the 1Inch and Paraswap solver. 

Currently, this component is unused, and will only get used in a follow up PR.

### Test Plan

Added unit tests for logic.