Make WeightedProductOrder implement BaselineSolvable (gnosis/gp-v2-services#795)
This PR implements `BaselineSolvable` for the new `WeightedProductOrder` so that the `Liquidity` type can be used directly by the baseline solver.

### Test Plan

Added some unit tests for the new BigRational <-> Bfp conversion logic. Otherwise, the rest was just "wiring" up implementations together and doesn't really need to be tested beyond compilation.