Disable pool intiailzation for unsupported networks (gnosis/gp-v2-services#810)
This PR adds a `DefaultPoolInitializer` type that automatically determines based on the network whether or not to use the subgrph Balancer pool intializer implementation. This is required because the Balancer subgraph requires an archive node in order to index pools with complete token data, which is only available on Mainnet for the Graph hosted service.

This means that for networks for which the Balancer subgraph doesn't index all pool data, we will need to index all events :cry:. Luckily, this can be somewhat mitigated by using the subgraph just to query pool addresses, and then reading the required pool data on-chain with `eth_call`s. This is done in a follow up PR.

### Test Plan

Added unit tests. Additionally, this code is used in #809 to make a settlement trading against Balancer pools.