Limit the gas for simulation transactions (gnosis/gp-v2-services#923)
This PR sets a gas limit for the simulation transaction of approximately the gas block limit. This is done because we now changed to specifying gas prices for simulations. Additionally, it appears that the node we use defaults to around 10x the block gas limit for `eth_call`s and return simulation failures for insufficient balance - even if the solver would actually have enough balance for the transaction.

### Test Plan

CI e2e test should cover this path (as it simulates the settlement).