Updated order struct hash computation for new contracts (gnosis/gp-v2-services#920)
Fixes #917

This PR updates the order struct hash computation to include the new balance fields. Updating this struct hash affects:
- Order UID computation
- Signature verification

Generated some new test cases base on the `gp-v2-contracts` `v1.0.0` release. Can be tested manually with `cargo test -p model`.

Note that CI is expected to fail because we are in a limbo state where part of the services are for the upgraded SCs, but they have not been enabled yet.