Keep single order solving until deadline runs out (gnosis/gp-v2-services#1011)
We used to fire off the solving process for N orders simultaneously with
each process retrying up to M times. Now we keep on solving the next order
until the deadline is reached without retries.
This has the advantage that the number of orders solved adapts to the
deadline so we always solve as many orders as possible.

Removed the parallelness of the previous code because it makes this
version more complex and it is nicer to the underlying solving apis. We
can reintroduce it later if we feel the need.

Finally I changed the tests of the inner single order solvers to not use
the outer SingleOrderSolving struct because it feels more appropriate to
me and I got rid of the method those tests were using.