Use Order Full Fee Amount In Solver Objective Value (gnosis/gp-v2-services#1204)
Fixes #1202

We now store full fee amount in the database, so we don't have to calculate it based on the fee factor.

The old calculation logic is preserved for the migration period. We will remove it after the next release, once there are no active orders with missing full fee amount.