Instrument price estimators for orderbook API (gnosis/gp-v2-services#1282)
This PR instruments (i.e. adds metrics for) price estimators that are used in the orderbook API. This will allow us to get a sense for which one is being used the most as well as measure cases where one starts failing a lot.

### Test Plan

Added unit test for newly instrumented price estimator.