Update to `tracing-subscriber v0.3.1` (gnosis/gp-v2-services#1302)
Supersedes #1293 

This PR updates the `tracing-subscriber` dependency to `v0.3.1`. Its a little more involved so I made a separate PR.

In particualr, time formatting has changed in the newer version to use the `time` crate, so the format string is specified differently. There is a built-in RFC 3339 which is almost what we need, except it uses a variable number of subsecond digits (based on trailing `0`s), so a custom format was needed.

### Test Plan

CI, no logic changes.

You can test that the formatting works by:
```
$ cargo run -p solver
2021-10-26T18:08:38.677Z  INFO solver: running solver with validated Arguments {
...
```

And then compare it to `main` to make sure its the same:
```
$ cargo run -p solver
2021-10-26T18:09:49.659Z  INFO solver: running solver with validated Arguments {
...
```